Handle scenarios where EndpointSlices are missing Zone
We saw a case where the xds servers were spinning because an
EndpointSlice/Address did not specify a Zone. We should still be able to
serve those endpoints to clients and create an alert to manually check
things.
Kube-controller-manager is responsible for populating the zone
in EndpointSlices.
We set the value to "none". This will not affect existing routing
algorithms but should be taken into account if we implement zone
specific routing strategies in the future.