Add a MaxConnectionIdle to xDS server
Since an idle_timeout was introduced by default in the client:
https://github.com/grpc/grpc-go/pull/6585/files we are observing clients
closing watches on xds resources after 30 minutes and when they are
trying to wake from IDLE state are not receiving proper updates from the
server. Most likely client and server are on different state where the server does
not think it should provide any update while the client has offloaded
all the Resolver resources and needs a complete LDS update (at least).
MaxConnectionIdle should force the server to send a `GoAway` after 30 minutes
and trigger the connection to be re-established if needed.