Refactor structure to be able to use the new kube::codegen helper funcs
Change structure and filenames to the ones expected by the codegen
functions to be able to find the CRD group and generate code. This is
the same structure a new kubebuilder generated project would have.
Also, stop using the deprecated (and also removed in the latest codegen
version) kube_codegen.sh script and start using the new way to generate
code.